+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add Epoch SwathGrid for swath sampler
+
 ### Added
 
 - Add StationSamplerMod for station sampler